toolbar moved inside media-viewer
