+++ projects/media-viewer/src/lib/media-viewer.module.ts
@@ -10,7 +10,7 @@ import { UnsupportedViewerComponent } from './media-viewer/viewers/unsupported-v
 import { MediaViewerComponent } from './media-viewer/media-viewer.component';
 import { MediaViewerService } from './media-viewer/media-viewer.service';
 import { EmLoggerService } from './logging/em-logger.service';
-import { ToolbarModule } from './toolbar/toolbar.module';
+import { ToolbarModule } from './media-viewer/toolbar/toolbar.module';
 
 @NgModule({
   imports: [
