+++ projects/media-viewer/src/lib/media-viewer/media-viewer.component.html
@@ -1,10 +1,19 @@
+<app-toolbar [actionEvents]="actionEvents" [url]="url" [contentType]="contentType"></app-toolbar>
+
 <app-pdf-viewer *ngIf="contentType === 'pdf'"
                           [url]="url"
                           [zoomOperation]="zoom | async"
                           [searchOperation]="search | async"
-                          [rotateOperation]="rotation | async"></app-pdf-viewer>
-<app-image-viewer *ngIf="contentType === 'image'" [url]="url" ></app-image-viewer>
-<app-unsupported-viewer *ngIf="contentTypeUnsupported()" ></app-unsupported-viewer>
+                          [rotateOperation]="rotation | async">
+</app-pdf-viewer>
+<app-image-viewer *ngIf="contentType === 'image'"
+                  [url]="url"
+                  [zoomOperation]="zoom | async"
+                  [rotateOperation]="rotation | async">
+</app-image-viewer>
+<app-unsupported-viewer [url]="url"
+                        *ngIf="contentTypeUnsupported()" >
+</app-unsupported-viewer>
 
 <div class="grid-row">
   <div *ngIf="error"  class="error-summary" role="alert" aria-labelledby="error-summary-heading-example-1" tabindex="-1">
