+++ projects/media-viewer/src/lib/media-viewer/toolbar/toolbar.module.ts
