+++ projects/media-viewer/src/lib/media-viewer/viewers/unsupported-viewer/unsupported-viewer.component.html
@@ -1,3 +1,3 @@
-<p data-hook="dm.viewer.unsupported">
-  This file type is not supported. Please <a href="{{url}}">click here to download</a>.
+<p data-hook="dm.viewer.unsupported" style="padding-top: 100px;">
+  This file type is not supported by this viewer. Please <a href="{{url}}">click here to download</a>.
 </p>
