+++ src/app/sandbox-webapp/sandbox-webapp.component.html
@@ -53,9 +53,7 @@
                    role="textbox"
                    aria-multiline="true" data-medium-editor-editor-index="1"
                    medium-editor-index="ca3cfb8f-8b6a-8f73-1ed5-fa7679703446" data-placeholder="Type your text">
-                <ng-container *ngIf="documentTypeToShow === 'pdf'">
-                  <app-toolbar [actionEvents]="actionEvents"></app-toolbar>
-                </ng-container>
+
                 <app-media-viewer *ngIf="documentTypeToShow === 'pdf'"
                                   [url]="'assets/example.pdf'"
                                   [actionEvents]=actionEvents
